feat(product): update and delete

+++ src/products/products.service.ts
@@ -8,11 +8,11 @@ import { UpdateProductDto } from './dto/update-product.dto';
 import { PrismaService } from '@/prisma/prisma.service';
 import { Query as ExpressQuery } from 'express-serve-static-core';
 import { Prisma, Product } from '@prisma/client';
-
+import { type } from 'os';
 
 @Injectable()
 export class ProductsService {
-  constructor(private prisma: PrismaService) {}
+  constructor(private prisma: PrismaService) { }
 
   async create(createProductDto: CreateProductDto, currentUser) {
     const { name, description, specifications, availableDays, rentalOptions } =
@@ -36,6 +36,9 @@ export class ProductsService {
           })),
         },
       },
+      include: {
+        rentalOptions: true,
+      }
     });
 
     return newProduct;
@@ -65,7 +68,6 @@ export class ProductsService {
     const query = await this.prisma.product.findMany({
       skip: skip,
       take: +perPage,
-
     });
     return query;
   }
@@ -79,29 +81,31 @@ export class ProductsService {
       updateProductDto;
 
     try {
-      // Check if the product exists
       const existingProduct = await this.prisma.product.findUnique({
         where: { id },
+        include: {
+          rentalOptions: true,
+        }
       });
 
       if (!existingProduct) {
         throw new NotFoundException('Product not found');
       }
 
-      // Check if the current user has permission to update this product
       if (existingProduct.ownerId !== currentUser.id) {
         throw new BadRequestException(
           'You do not have permission to update this product',
         );
       }
 
-      // Update the product
       const updatedProduct = await this.prisma.product.update({
         where: { id },
         data: {
           name,
           description,
-          specifications,
+          specifications: {
+            ...specifications,
+          },
           availableDays: {
             update: {
               startDate: new Date(availableDays.startDate),
@@ -109,21 +113,36 @@ export class ProductsService {
             },
           },
           availability: true,
-          rentalOptions: {
-            upsert: rentalOptions.map((option) => ({
-              where: { type: option.type },
-              create: {
-                type: option.type,
-                priceRate: option.priceRate,
-              },
-              update: {
-                type: option.type,
-                priceRate: option.priceRate,
-              },
-            })),
-          },
-        },
-      });
+        }
+      })
+
+      for (const rentalOption of rentalOptions) {
+        const existingRental = await existingProduct.rentalOptions.find((rental) => rental.type === rentalOption.type)
+        if (existingRental) {
+          if (rentalOption.isSelected) {
+            await this.prisma.rentalOption.update({
+              where: { id: existingRental.id },
+              data: {
+                priceRate: rentalOption.priceRate
+              }
+            });
+          }
+          else {
+            await this.prisma.rentalOption.delete({
+              where: { id: existingRental.id },
+            });
+          }
+        }
+        else if (rentalOption.isSelected) {
+          await this.prisma.rentalOption.create({
+            data: {
+              productId: existingProduct.id,
+              type: rentalOption.type,
+              priceRate: rentalOption.priceRate,
+            },
+          });
+        }
+      }
 
       return updatedProduct;
     } catch (error) {
@@ -131,7 +150,32 @@ export class ProductsService {
     }
   }
 
-  remove(id: string) {
-    return this.prisma.product.delete({ where: { id } });
+
+  async remove(id: string, currentUser) {
+    try {
+      const existingProduct = await this.prisma.product.findUnique({
+        where: { id },
+        include: {
+          rentalOptions: true,
+        }
+      });
+
+      if (!existingProduct) {
+        throw new NotFoundException('Product not found');
+      }
+
+      if (existingProduct.ownerId !== currentUser.id) {
+        throw new BadRequestException(
+          'You do not have permission to update this product',
+        );
+      }
+      await this.prisma.rentalOption.deleteMany({ where: { productId: id } });
+      await this.prisma.product.delete({ where: { id } });
+
+      return { message: 'Product deleted successfully' };
+    } catch (error) {
+      throw new BadRequestException('Failed to delete product');
+    }
+
   }
 }
